+++ src/main/java/org/seedstack/maven/EffectiveTestConfigMojo.java
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.seedstack.maven;
+
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.plugins.annotations.Execute;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.ResolutionScope;
+import org.seedstack.maven.runnables.ToolLauncherRunnable;
+
+/**
+ * Defines the effective-test-config goal. This goal runs the effective-config Seed tool with the test classpath,
+ * which dumps the effective test configuration.
+ */
+@Mojo(name = "effective-test-config", requiresProject = true, threadSafe = true, defaultPhase = LifecyclePhase.VALIDATE, requiresDependencyResolution = ResolutionScope.TEST)
+@Execute(phase = LifecyclePhase.PROCESS_TEST_CLASSES)
+public class EffectiveTestConfigMojo extends AbstractExecutableMojo {
+    public EffectiveTestConfigMojo() {
+        enableTestMode();
+    }
+
+    @Override
+    public void execute() throws MojoExecutionException, MojoFailureException {
+        runnable = new ToolLauncherRunnable("effective-config", getArgs(), getMonitor(), getLog());
+        super.execute();
+    }
+}